[MAAS-588] Add introduction to the FRED tutorials repo
We reuse much of the text from the introduction to the 'FRED Tutorial
Library' chapter from the FRED Guide [1] to provide a introduction to
the content and intended audience of the tutorials repository. We also
provide a contact email address so readers can request access to the
guide if they don't have it. As discussed on Slack [2] this will be
handled by Erik via the support@epistemix.com account.

[1]:
https://epistemix-fred-guide.readthedocs-hosted.com/en/dev/user_guide/chapter21/chapter21.html#chapter-21-the-fred-tutorial-library
[2]: https://epistemix.slack.com/archives/C01FCU4GNGH/p1612963838002800?thread_ts=1612960923.002400&cid=C01FCU4GNGH